Otras IA/MEA-Core-AI-System_v2.0.0_20250809_211629/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.3.2",
    "pdfminer-six>=20250506",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
]
